chore: update /snapshots .p mirror
